+++ tests/integration/test_index_routines.py
@@ -19,7 +19,7 @@ from itertools import permutations
 import numpy as np
 import pytest
 from cunumeric.eager import diagonal_reference
-from test_tools.generators import mk_seq_array
+from utils.generators import mk_seq_array
 
 import cunumeric as num
 from legate.core import LEGATE_MAX_DIM
