+++ tests/integration/test_norm.py
@@ -15,7 +15,7 @@
 
 import numpy as np
 import pytest
-from test_tools.generators import mk_0to1_array
+from utils.generators import mk_0to1_array
 
 import cunumeric as cn
 from legate.core import LEGATE_MAX_DIM
