+++ tests/integration/test_singleton_access.py
@@ -15,7 +15,7 @@
 
 import numpy as np
 import pytest
-from test_tools.generators import mk_0to1_array, scalar_gen
+from utils.generators import mk_0to1_array, scalar_gen
 
 import cunumeric as num
 from legate.core import LEGATE_MAX_DIM
