+++ tests/integration/test_trace.py
@@ -18,7 +18,7 @@ from itertools import permutations
 
 import numpy as np
 import pytest
-from test_tools.generators import mk_seq_array
+from utils.generators import mk_seq_array
 
 import cunumeric as num
 from legate.core import LEGATE_MAX_DIM
